+++ frontend/src/App/App.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import App from "./App";
+import App from "./index";
 
 test("renders test element in App", () => {
   render(<App />);
